Rename "misc" application to "cms".
The miscellaneous pieces of code in the "misc" application all relate to
content management.
