+++ conf_site/cms/apps.py
@@ -1,0 +1,7 @@
+from __future__ import unicode_literals
+
+from django.apps import AppConfig
+
+
+class CmsConfig(AppConfig):
+    name = 'cms'
